S7 : fix decodeAddress()

+++ src/javaforce/controls/s7/S7Data.java
@@ -7,10 +7,10 @@ package javaforce.controls.s7;
 
 public class S7Data {
   public byte block_type;
-  public short block_number;
+  public int block_number;
   public byte data_type;
   public int offset;  //24bit
-  public short length;  //# elements
+  public int length;  //# elements
   public byte data[];
   public int getLength() {  //returns bytes
     return S7Types.getTypeSize(data_type, length);
